Move ValueSetter out of the Chunk class
